user and event components

+++ src/app/private/components/user-list/user-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { User } from 'interfaces/user';
-import { UserService } from 'services/userService/user.service';
+import { Router } from '@angular/router';
+import { UserService } from './userService/user.service';
 
 @Component({
   selector: 'app-user-list',
@@ -9,50 +10,59 @@ import { UserService } from 'services/userService/user.service';
 })
 export class UserListComponent implements OnInit {
 
-  onDelete:boolean
+  confirmDeleteId:number
   users: User[] = [];
 
-  constructor(private userService: UserService) {
-    this.onDelete = false
+  constructor(private userService: UserService, private router: Router) {
+    this.confirmDeleteId = 0;
   }
 
   ngOnInit(): void {
     this.getUsers();
   }
 
+  redirectToNewUsertPage(){
+    this.router.navigate(['/private/addUser']);
+  }
+
+  redirectToEditPage(userId: number | undefined) {
+    console.log("userId", userId);
+    if(userId !== undefined) {
+      this.router.navigate(['/private/editUser', userId]);
+    }
+  }
 
   private getUsers(){
-    this.userService.getUserList().subscribe(data => {
-      this.users = data;
+    let filter = {
+      "role": "MEMBER"
+    }
+    this.userService.getUserList(filter).subscribe(data => {
+      console.log('data', data.content);
+      this.users = data.content;
     })
   }
 
-  onEnableUser(userId: number) {
-    this.userService.enableUser(userId).subscribe(
-      (response) => {
-        if(response == "This Account enabled with success !!!!!")
-        window.location.reload();
-      }
-    )
-  }
-
-  onDisableUser(userId: number) {
-    this.userService.disableUser(userId).subscribe(
-      (response) => {
-        if(response == "This Account is disabled !!!!!")
-        window.location.reload();
-      }
-    )
-  }
-
-  onDeleteAccount(userId: number) {
-    this.userService.deleteUser(userId).subscribe(
-      (response) => {
-        if(response == "this Account is deleted"){
-          window.location.reload();
+  onEnableUser(userId: number | undefined) {
+    if(userId !== undefined) {
+      this.userService.enableUser(userId).subscribe(
+        (response) => {
+          if(response == "This Account enabled with success !!!!!")
+              this.getUsers();
         }
-      }
-    )
+      )
+    }
+  }
+
+  onDeleteAccount(userId: number | undefined) {
+    if(userId !== undefined) {
+      this.userService.deleteUser(userId).subscribe(
+        (response) => {
+          if(response == "this Account is deleted"){
+            this.getUsers();
+          }
+        }
+      )
+    }
   }
 
 }
